backend/pyproject.toml
[project]
name = "venv"
version = "1.0.0"
description = "AI-Powered Resume Analyzer Backend"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "python-multipart==0.0.6",
    "PyPDF2==3.0.1",
    "python-dotenv==1.0.0",
    "google-generativeai>=0.3.0",
    "markdown>=3.5.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "httpx>=0.24.0",
]
